Fix TeXLive 2019 not building
After recent release of TeXLive 2020, the approach of using
install-tl-unx.tar.gz for installing LaTeX was running smoothly for 2017
and 2018 versions.

But all of a sudden stopped working for 2019. This was the error I
received:

./install-tl: unexpected return value from verify_checksum: -5

Debugging into the problem, I could not find help. Alternatively I opted
for use of install-tl.zip file. With its use, its working great.